requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-postgres
pyxlsb==1.0.10
pandas==2.2
openpyxl==3.1.2
xlsxwriter==3.2.0
grpcio-status==1.62.1